+++ test/base/src/androidTest/kotlin/io/realm/test/shared/RealmTests.kt
@@ -20,6 +20,7 @@ import io.realm.RealmConfiguration
 import io.realm.VersionId
 import io.realm.entities.link.Child
 import io.realm.entities.link.Parent
+import io.realm.internal.RealmConfigurationImpl
 import io.realm.internal.interop.NativePointer
 import io.realm.internal.platform.WeakReference
 import io.realm.isManaged
@@ -397,19 +398,69 @@ class RealmTests {
     }
 
     @Test
-    fun closingIntermediateVersionsWhenNoLongerReferenced() {
+    fun intermediateVersionsReleaseWhenProgressingRealm() {
         assertEquals(0, intermediateReferences.value.size)
-        var parent: Parent? = realm.writeBlocking { copyToRealm(Parent()) }
+        realm.writeBlocking { }
         assertEquals(1, intermediateReferences.value.size)
         realm.writeBlocking { }
+        assertEquals(2, intermediateReferences.value.size)
+        realm.writeBlocking { }
+        assertEquals(3, intermediateReferences.value.size)
+
+        // Trigger GC - On native we also need to trigger GC on the background thread that creates
+        // the references
+        runBlocking((realm.configuration as RealmConfigurationImpl).writeDispatcher) {
+            triggerGC()
+        }
+        triggerGC()
+
+        // Close of intermediate version is currently only done when updating the realm after a write
+        realm.writeBlocking { }
+        assertEquals(1, intermediateReferences.value.size)
+    }
+
+    @Test
+    // TODO Investigate why clearing local object variable does not trigger collection of
+    //  reference on Native. Could just be that the GC somehow does not collect this when
+    //  cleared due some thresholds or outcome of GC not being predictable.
+    //  https://github.com/realm/realm-kotlin/issues/486
+    @Ignore
+    fun clearingRealmObjectReleasesRealmReference() {
+        assertEquals(0, intermediateReferences.value.size)
+        // The below code creates the object without returning it from write to show that the
+        // issue is not bound to the freezing inside write, but also happens on the same thread as
+        // the realm is constructed on.
+        realm.writeBlocking { copyToRealm(Parent()); Unit }
+        var parent: Parent? = realm.objects<Parent>()!!.first()
+        assertEquals(1, intermediateReferences.value.size)
+        realm.writeBlocking { }
+        assertEquals(2, intermediateReferences.value.size)
+        realm.writeBlocking { }
+        assertEquals(3, intermediateReferences.value.size)
+
+        // Trigger GC - On native we also need to trigger GC on the background thread that creates
+        // the references
+        runBlocking((realm.configuration as RealmConfigurationImpl).writeDispatcher) {
+            triggerGC()
+        }
+        triggerGC()
+
+        // Close of intermediate version is currently only done when updating the realm after a write
+        realm.writeBlocking { }
+        // We still have the single intermediate reference as a result of the write itself
+        // and the reference kept alive by the realm object
         assertEquals(2, intermediateReferences.value.size)
 
         // Clear reference
         parent = null
-        // Trigger GC
+
+        runBlocking((realm.configuration as RealmConfigurationImpl).writeDispatcher) {
+            triggerGC()
+        }
         triggerGC()
-        // Close of intermediate version is currently only done when updating the realm after a write
+
         realm.writeBlocking { }
+        // Clearing the realm object reference allowed clearing the corresponding reference
         assertEquals(1, intermediateReferences.value.size)
     }